mise a jour

+++ src/main/java/fr/isika/projet3/controller/AssociationController.java
@@ -119,7 +119,7 @@ public class AssociationController {
 
     //edit one Association
     @RequestMapping(value = "/editAssociation/{id}", method = RequestMethod.GET)
-    public ModelAndView displayEditAssociationForm(@PathVariable int id) {
+    public ModelAndView displayEditAssociationForm(@PathVariable Long id) {
         ModelAndView mv = new ModelAndView("/editAssociation");
         Association association = associationService.getAssociationById(id);
         mv.addObject("headerMessage", "Edit Association Details");
@@ -146,7 +146,7 @@ public class AssociationController {
 
     //delete one association
     @RequestMapping(value = "/deleteAssociation/{id}", method = RequestMethod.GET)
-    public ModelAndView deleteAssociationById(@PathVariable int id) {
+    public ModelAndView deleteAssociationById(@PathVariable Long id) {
         boolean isDeleted = associationService.deleteAssociationById(id);
         System.out.println("Association deletion response: " + isDeleted);
         ModelAndView mv = new ModelAndView("redirect:/home");
@@ -157,7 +157,7 @@ public class AssociationController {
 
     //edit one Association
     @RequestMapping(value = "/homePageAssociation/{id}", method = RequestMethod.GET)
-    public ModelAndView displayHomePageAssociationForm(@PathVariable int id) {
+    public ModelAndView displayHomePageAssociationForm(@PathVariable Long id) {
         ModelAndView mv = new ModelAndView("/homePageAssociation");
         Association association = associationService.getAssociationById(id);
         mv.addObject("headerMessage", "Edit home Page Association Details");
